Move modal to components/modals 🔥
